fix: slice init was incorrect => there were null pointers in the list of the sub's and collector had panic on null pointers + minor changes with linters, but i think them to be incorrect anyway

+++ task4/repo-stat/subscribe/internal/usecase/db_usecase.go
@@ -8,10 +8,11 @@ import (
 
 type DBUseCase struct {
 	retriever domain.Retriever
+	pinger    *PingUseCase
 }
 
-func NewDBUseCase(retriever domain.Retriever) *DBUseCase {
-	return &DBUseCase{retriever: retriever}
+func NewDBUseCase(retriever domain.Retriever, pinger *PingUseCase) *DBUseCase {
+	return &DBUseCase{retriever: retriever, pinger: pinger}
 }
 func (uc *DBUseCase) List(ctx context.Context) ([]domain.Slug, error) {
 	return uc.retriever.List(ctx)
@@ -22,12 +23,29 @@ func (uc *DBUseCase) Push(ctx context.Context, slug domain.Slug) error {
 		return fmt.Errorf("failed to list subs: %v", err)
 	}
 	for _, el := range slugs {
-		if el == slug {
-			return fmt.Errorf("failed to push the sub %s/%s, dublication error", slug.Owner, slug.Repo)
+		if el.Owner == slug.Owner && el.Repo == slug.Repo {
+			return fmt.Errorf("failed to push the sub %s/%s: 409: dublication error", slug.Owner, slug.Repo)
 		}
+	}
+	err = uc.pinger.PingRepo(ctx, slug)
+	if err != nil {
+		return fmt.Errorf("failed to push the sub %s/%s: %v", slug.Owner, slug.Repo, err)
 	}
 	return uc.retriever.Push(ctx, slug)
 }
 func (uc *DBUseCase) Delete(ctx context.Context, slug domain.Slug) error {
+	slugs, err := uc.retriever.List(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to list subs: %v", err)
+	}
+	var flag bool = false
+	for _, el := range slugs {
+		if el.Owner == slug.Owner && el.Repo == slug.Repo {
+			flag = true
+		}
+	}
+	if !flag {
+		return fmt.Errorf("failed to delete the sub %s/%s: %v", slug.Owner, slug.Repo, "404: subscription not found")
+	}
 	return uc.retriever.Delete(ctx, slug)
 }
